% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmvnormAR1.R
\name{dmvnormAR1}
\alias{dmvnormAR1}
\title{loglikelihood for AR1 model}
\usage{
dmvnormAR1(y, phi, sigmaAR1)
}
\arguments{
\item{y}{the data}

\item{phi}{the autocorrelation parameter}

\item{sigmaAR1}{the variance parameter}
}
\value{
the loglikelihood
}
\description{
evaluates to the loglikelihood for an AR1 model conditional on the first value being 0
}
\author{
Jay Ver Hoef
}
