% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haulout_initial_JHop.R
\name{haulout_initial_JHop}
\alias{haulout_initial_JHop}
\title{Initialize haulout MCMC chain}
\usage{
haulout_initial_JHop(season)
}
\arguments{
\item{season}{either "spring" or "fall".}
}
\value{
a list of inputs for stock_haulout_estimation() function
}
\description{
Initialize haulout MCMC chain
}
\author{
Jay Ver Hoef
}
