% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL.R
\name{LL}
\alias{LL}
\title{loglikelihood for model}
\usage{
LL(eta, tau, xi, tauxi, ppi, phiYr, sigmaYr, phiDyPup, sigmaDyPup, phiDyMolt,
  sigmaDyMolt, nu, Ni, WdiPup, WdiMolt, sumWdiPup, sumWdiMolt, Xpg, Ypg, Xpa,
  Ypa, Ymg, Yma, pdayrMin, mdayrMin)
}
\arguments{
\item{eta}{overall intercept}

\item{tau}{effect for survey method being ground count}

\item{xi}{effect for surveys in the pupping period}

\item{tauxi}{ground count by pupping period interaction}

\item{ppi}{effect for pup counts}

\item{phiYr}{AR1 autocorrelation parameter for yearly time series}

\item{sigmaYr}{AR1 variance parameter for yearly time series}

\item{phiDyPup}{AR1 autocorrelation parameter for daily time series during pupping period}

\item{sigmaDyPup}{AR1 variance parameter for daily time series during pupping period}

\item{phiDyMolt}{AR1 autocorrelation parameter for daily time series during molting period}

\item{sigmaDyMolt}{variance parameter for daily time series during molting period}

\item{Ni}{yearly time series random effects}

\item{WdiPup}{daily time series random effects during pupping period}

\item{WdiMolt}{daily time series random effects during molting period}

\item{sumWdiPup}{sum of the likelihood of daily random effects during pupping period}

\item{sumWdiMolt}{sum of the likelihood of daily random effects during the molting period}

\item{Xpg}{dataset for pup ground counts during pupping period}

\item{Ypg}{dataset for nonpup ground counts during pupping period}

\item{Xpa}{dataset for pup aerial survey estimates during pupping period}

\item{Ypa}{dataset for nonpup aerial survey estimates during pupping period}

\item{Ymg}{dataset for nonpup ground counts during molting period}

\item{Yma}{dataset for nonpup aerial survey estimates during molting period}

\item{pdayrMin}{minimum yday for all data sets during pupping period}

\item{mdayrMin}{minimum yday for all data sets during molting period}

\item{sigmaZ}{variance parameter for independent residuals}
}
\value{
the loglikelihood
}
\description{
evaluates to the loglikelihood for MCMC sampling
}
\author{
Jay Ver Hoef
}
