% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{MCMC}
\alias{MCMC}
\title{MCMC sampling}
\usage{
MCMC(burnin = 1, niter = 1000, thin = 2, eta, Ni, tau, xi, tauxi, ppi,
  phiYr, sigmaYr, phiDyPup, sigmaDyPup, phiDyMolt, sigmaDyMolt, nu, WdiPup,
  WdiMolt, Xpg, Ypg, Xpa, Ypa, Ymg, Yma)
}
\arguments{
\item{burnin}{the number of MCMC for burnin}

\item{niter}{the number of MCMC iterations}

\item{thin}{thinning of the MCMC chain}

\item{eta}{overall intercept}

\item{Ni}{yearly time series random effects}

\item{tau}{effect for survey method being ground count}

\item{xi}{effect for surveys in the pupping period}

\item{tauxi}{ground count by pupping period interaction}

\item{ppi}{effect for pup counts}

\item{phiYr}{AR1 autocorrelation parameter for yearly time series}

\item{sigmaYr}{AR1 variance parameter for yearly time series}

\item{phiDyPup}{AR1 autocorrelation parameter for daily time series during pupping period}

\item{sigmaDyPup}{AR1 variance parameter for daily time series during pupping period}

\item{phiDyMolt}{AR1 autocorrelation parameter for daily time series during molting period}

\item{sigmaDyMolt}{variance parameter for daily time series during molting period}

\item{nu}{coefficient of variation for lognormal model}

\item{WdiPup}{daily time series random effects during pupping period}

\item{WdiMolt}{daily time series random effects during molting period}

\item{Xpg}{dataset for pup ground counts during pupping period}

\item{Ypg}{dataset for nonpup ground counts during pupping period}

\item{Xpa}{dataset for pup aerial survey estimates during pupping period}

\item{Ypa}{dataset for nonpup aerial survey estimates during pupping period}

\item{Ymg}{dataset for nonpup ground counts during molting period}

\item{Yma}{dataset for nonpup aerial survey estimates during molting period}
}
\value{
a list of the MCMC chains values
}
\description{
MCMC sampling
}
\author{
Jay Ver Hoef
}
